Change type structure for LUIS entities
Renamed `BasicLuisEntity` interface -> `LuisEntityBase` and introduced
new union type definition under the original name made up of
`LuisEntity|LuisCompositeEntity`. This allows for type guarding rules
around the `type` property to naturally kick in.

Also created explicit type guard functions for `LuisEntity`
and `LuisCompositeEntity`.

In the process of this work, it became clear it was time to split the
code into smaller, more managable source files.
